+++ lib/validation.tsx
@@ -3,7 +3,6 @@ import {
   FormValues,
   DynamicFormProps,
   InnerFormProps,
-  FormElement,
   Responses,
   ValidationProperties,
 } from "./types";
@@ -75,6 +74,53 @@ const scrollErrorInView = (e: React.MouseEvent<HTMLAnchorElement, MouseEvent>, i
   }
 };
 
+const isFielddResponseValid = (
+  value: unknown,
+  componentType: string,
+  validator: ValidationProperties,
+  t: TFunction
+): string | null => {
+  switch (componentType) {
+    case "textField": {
+      const typedValue = value as string;
+      if (validator.required && !typedValue) return t("input-validation.required");
+      const currentRegex = getRegexByType(validator.type, t, value as string);
+      if (validator.type && currentRegex && currentRegex.regex) {
+        // Check for different types of fields, email, date, number, custom etc
+        const regex = new RegExp(currentRegex.regex);
+        if (typedValue && !regex.test(typedValue)) {
+          return currentRegex.error;
+        }
+      }
+      break;
+    }
+    case "textArea": {
+      const typedValue = value as string;
+      if (validator.required && !typedValue) return t("input-validation.required");
+      break;
+    }
+    case "checkbox":
+    case "radio":
+    case "dropdown": {
+      if (validator.required && value === undefined) return t("input-validation.required");
+      break;
+    }
+    case "fileInput": {
+      const typedValue = value as Record<string, string>;
+      if (validator.required && typedValue["file"] === null) return t("input-validation.required");
+      break;
+    }
+    case "dynamicRow": {
+      // Not implemented yet
+      break;
+    }
+    default:
+      throw `Validation for component ${componentType} is not handled`;
+  }
+
+  return null;
+};
+
 /**
  * validateOnSubmit is called during Formik's submission event to validate the fields
  * @param values
@@ -84,37 +130,20 @@ export const validateOnSubmit = (values: FormValues, props: DynamicFormProps): R
   const errors: Responses = {};
 
   for (const item in values) {
-    const formConfig = props.formConfig;
-    const elements: Array<FormElement> = formConfig.elements;
-    const currentItem = elements.find((element) => element.id == item);
+    const formElement = props.formConfig.elements.find((element) => element.id == item);
 
-    if (!currentItem) {
-      return errors;
-    }
+    if (!formElement) return errors;
 
-    const currentValidation = (currentItem.properties?.validation as ValidationProperties) || {};
-    const formikValue = values[item];
-    // A non-array object indicates an unset field (this is how Formik sends it to us)
-    // Discard these values, and convert anything else to a string
-    const currentValue: string = (
-      typeof formikValue == "object" && !Array.isArray(formikValue) ? "" : (formikValue as string)
-    ).toString();
-    const currentRegex = getRegexByType(currentValidation.type, props.t, currentValue);
+    if (formElement.properties.validation) {
+      const result = isFielddResponseValid(
+        values[item],
+        formElement.type,
+        formElement.properties.validation,
+        props.t
+      );
 
-    // Check for required fields
-    let isEmpty = !currentValue;
-    // currentValue is not a straight existence check for things like radio button and checkboxes
-    if (!isEmpty && typeof formikValue == "object") {
-      if (Array.isArray(formikValue) && formikValue.length < 1) isEmpty = true;
-      if (!Array.isArray(formikValue)) isEmpty = true; // non array objects are an unset field
-    }
-    if (currentValidation && currentValidation?.required && isEmpty) {
-      errors[item] = props.t("input-validation.required");
-    } else if (currentValidation.type && currentRegex && currentRegex.regex) {
-      // Check for different types of fields, email, date, number, custom etc
-      const regex = new RegExp(currentRegex.regex);
-      if (currentValue && !regex.test(currentValue)) {
-        errors[item] = currentRegex.error;
+      if (result) {
+        errors[item] = result;
       }
     }
   }